requirements.txt
openai-whisper
pydub

